refactor(libraries): move TransformableImage & ViewPager in separate folders
